requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Cors==3.0.10
Flask-Limiter==2.8.0
Flask-Migrate==4.0.4
SQLAlchemy==2.0.15
psycopg2-binary==2.9.6
python-dotenv==1.0.0
dnspython==2.3.0
gunicorn==20.1.0
psutil==5.9.4
Werkzeug==2.2.3
email-validator==1.3.1
ipaddress==1.0.23
requests>=2.25.1
beautifulsoup4>=4.9.3
cryptography>=3.4.7
uuid>=1.30
python-dateutil>=2.8.1
Jinja2>=3.0.0
itsdangerous>=2.0.0
click>=8.0.0
Pillow>=8.0.0
stripe>=3.0.0
cssutils>=2.3.0